+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fix memory leaks in integrations. (PR #4)
     * Integrate improvements to EOO/callsign reporting. (PR #4, #5)
     * Flex: Enable SNR meter even if reporting is disabled. (PR #6)
+    * Flex: Only go into TX if the client with the FDV slice is transmitting. (PR #12)
 2. Feature enhancements:
     * Enable UDP broadcast of received callsigns if FreeDV Reporter is enabled. (PR #1, #8)
     * Flex: allow some command line arguments to be passed via the environment. (PR #5)